Moving AppConfig out of attraction, with database in utils
