+++ code/src/main/java/utils/AppConfig.java
@@ -1,4 +1,4 @@
-package attraction;
+package utils;
 import jakarta.ws.rs.ApplicationPath;
 import jakarta.ws.rs.core.Application;
