+++ code/src/main/java/utils/DBManager.java
@@ -1,7 +1,8 @@
 /**
  * 
  */
-package database;
+
+package utils;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
